Refactored AgentNode for v2
