+++ packages/@n8n/nodes-langchain/nodes/agents/Agent/agents/ToolsAgent/V1/description.ts
@@ -1,0 +1,19 @@
+import type { INodeProperties } from 'n8n-workflow';
+
+import { commonOptions } from '../options';
+
+export const toolsAgentProperties: INodeProperties[] = [
+	{
+		displayName: 'Options',
+		name: 'options',
+		type: 'collection',
+		displayOptions: {
+			show: {
+				agent: ['toolsAgent'],
+			},
+		},
+		default: {},
+		placeholder: 'Add Option',
+		options: [...commonOptions],
+	},
+];
